+++ lib/src/native.dart
@@ -3,7 +3,8 @@ part of '../flutter_curiosity.dart';
 typedef AndroidActivityResultHandler = void Function(
     AndroidActivityResult result);
 
-typedef KeyboardStatusHandler = void Function(bool visibility);
+typedef NativeKeyboardStatusCallback = void Function(
+    NativeKeyboardStatus status);
 
 class NativeTools {
   factory NativeTools() => _singleton ??= NativeTools._();
@@ -11,15 +12,14 @@ class NativeTools {
   NativeTools._() {
     _channel.setMethodCallHandler((MethodCall call) async {
       switch (call.method) {
-        case 'keyboardStatus':
-          for (var element in keyboardStatus) {
-            element(call.arguments as bool);
+        case 'onKeyboardStatus':
+          for (var element in _keyboardStatus) {
+            element(NativeKeyboardStatus.formMap(call.arguments as Map));
           }
           break;
         case 'onActivityResult':
           for (var element in activityResult) {
-            element.call(AndroidActivityResult.formJson(
-                call.arguments as Map<dynamic, dynamic>));
+            element.call(AndroidActivityResult.formMap(call.arguments as Map));
           }
           break;
       }
@@ -28,11 +28,33 @@ class NativeTools {
 
   static NativeTools? _singleton;
 
-  /// 添加回调方法
-  final List<AndroidActivityResultHandler> activityResult = [];
+  /// 添加键盘回调方法
+  final List<NativeKeyboardStatusCallback> _keyboardStatus = [];
+
+  /// add Keyboard listener
+  Future<bool> addKeyboardListener(
+      NativeKeyboardStatusCallback callback) async {
+    if (!_supportPlatform) return false;
+    final needCallNative = _keyboardStatus.isEmpty;
+    if (!_keyboardStatus.contains(callback)) {
+      _keyboardStatus.add(callback);
+    }
+    if (!needCallNative) return true;
+    return (await _channel.invokeMethod<bool>('addKeyboardListener')) ?? false;
+  }
+
+  /// remove Keyboard listener
+  Future<bool> removeKeyboardListener(
+      NativeKeyboardStatusCallback callback) async {
+    if (!_supportPlatform) return false;
+    _keyboardStatus.remove(callback);
+    if (_keyboardStatus.isNotEmpty) return true;
+    return (await _channel.invokeMethod<bool>('removeKeyboardListener')) ??
+        false;
+  }
 
   /// 添加回调方法
-  final List<KeyboardStatusHandler> keyboardStatus = [];
+  final List<AndroidActivityResultHandler> activityResult = [];
 
   /// 判断GPS是否开启，GPS或者AGPS开启一个就认为是开启的
   Future<bool> get gpsStatus async {
@@ -55,7 +77,7 @@ class NativeTools {
     final List<AppPackageInfo> list = [];
     if (appList != null) {
       for (final dynamic data in appList) {
-        list.add(AppPackageInfo.fromJson(data as Map<dynamic, dynamic>));
+        list.add(AppPackageInfo.formMap(data as Map));
       }
     }
     return list;